backend/requirements.txt
# BE-00 — Pinned to major.minor for stability. No new deps without approval.
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
supabase>=2.3.0,<3.0
pydantic-settings>=2.0.0,<3.0
python-dotenv>=1.0.0,<2.0
PyJWT>=2.8.0,<3.0
httpx>=0.27.0,<1.0
openai>=1.12.0,<2.0
# IN-00: Fernet for connector secrets encryption
cryptography>=42.0.0,<43.0
stripe>=14.4.1,<15.0